Consistent decodings for utf8
Passing encoded byte arrays to `str` in python3 causes them to serialize to `"b'my byte array'"`, which... is not what we want. This causes authentication errors when we attempt to pass the funky serialization for the access token.

Added debug flag

This includes a --debug flag for debugging. Also prints connection information for github, _including the access token_, which is needed to debug an authentication issue.