Adds models for configuration
Since configuration is getting a little complicated; let's do all the parsing up front. This defines a grammar to convert from yaml to python objects. Now we can use it as an interop layer to read/write configs to/from whatever syntax we like. If we wanted to support other formats it makes for an easy entry point.

This also allows for logic to override configurations easily, and stricter, more consistent definitions.