Adds regex support
If you want to watch for a regular expression change in either a source or a target file; you can specify `watched_regexes` in your `~/.github-watcher.yml` file.

Those regexes will be applied to all diffs corresponding to open pull requests, even if those files are not in your watched files/directories in the main config.